backend/requirements.txt
flask
openai
azure-identity 
azure-keyvault-secrets
requests

